Draft: move the TestDraft unit tests to separate modules
This results in better organization of the tests,
avoiding extremely big files.
