+++ src/Mod/Draft/drafttests/__init__.py
@@ -1,0 +1,1 @@
+#
